Add ceph-disk to build
This needs to be done because we want to keep compatibility with
puppet-ceph-2.4.1-1 and our current version of puppet.

As this version of puppet-ceph only uses ceph-disk we will keep it
until we are able to move on to ceph-volume. Probably this will be
possible when startlingx is using version 3.1.1 of puppet-ceph.

Story: 2009074
Task: 43465

Signed-off-by: Delfino Curado <delfinogomes.curadofilho@windriver.com>
Change-Id: Ie9570f01728df28ee4ea357b1e618c5a4c0a3803