Change ceph-mon configuration
Now the "auth allow insecure global id reclaim" is false by default to
all monitors. By default ceph nautilus keeps this enabled to make possible the
cluster have a mixed set of ceph versions.

As this is not the case during an initial installation, we can add
this by default.

Gdisk was added again as this is necessary for StarlingX

Story: 2009074
Task: 43464

Signed-off-by: Delfino Curado <delfinogomes.curadofilho@windriver.com>
Change-Id: I5f3e432444b60ab73136431bb94bb6ab532ae0ab