Add debian package for puppet-network
This package is 3rdparty. Get latest release that supports Puppet 5,
which is 0.9.0. Get few commits after, such that requirements better
align with other packages we provide (puppet modules:
stdlib 5.x, boolean 2.x).
The commit has the module version of 0.9.1-rc0.

Considering patches in alphabetical order.
Patches 1, 2 are not needed as the changes are present upstream.
Patch 3 not needed as we don't care about the redhat change in
it and the rest of the changes are merged upstream.
Didn't port patch 4, not sure we need it. Assume we don't need it.
From patch 5 dropped WRlinux OS support.

TODO: Port gateway property change from patch 5. Port patches 6 and 7.

Did build-pkgs. Did build iso.

Story: 2009101
Task: 43359
Signed-off-by: Dan Voiculeasa <dan.voiculeasa@windriver.com>
Change-Id: Id94e43f1a556c83e4ffd429e0937a80258728b02