Add debian support for puppet-ceph
Add debian infrastructure to build puppet-ceph as a debian packaging.

Patches Dropped:

0005-Remove-puppetlabs-apt-as-ceph-requirement.patch
0004-US92424-Add-OSD-support-for-persistent-naming.patch
0006-ceph-disk-prepare-invalid-data-disk-value.patch
0008-ceph-mimic-prepare-activate-os.patch
0009-fix-ceph-osd-disk-partition-for-nvme-disks.patch

Re-diffed:

0001-Roll-up-TIS-patches.patch
0002-Newton-rebase-fixes.patch
0003-ceph-jewel-rebase.patch
0004-US92424-Add-OSD-support-for-persistent-naming.patch
0005-Add-StarlingX-specific-restart-command-for-Ceph-moni.patch

The patches that were dropped needs to be re-worked for ceph-volume,
since ceph-disk has been deprecated and not included in the Ceph version
from Debian.

0001-Roll-up-TIS-patches.patch and 0003-ceph-jewel-rebase.patch
were rebased because most of the systemd logic has
been dropped as well and need to be redone.

Story: 2009101
Task: 43431

Signed-off-by: Charles Short <charles.short@windriver.com>
Change-Id: I5a66bcd274f2752d4c050fab25a7b1b8347b650e