Add toleration to ansible-api deployment
A toleration needs to be added to all resources that create pods since
the node-role.kubernetes.io/master taint will be restored to all master
nodes. This ensures that the pods will run on the master node.

Test plan:

- Building and installing a new iso with this patch and the taint
applied. The armada-api pod was able to run on the tainted node.
- Confirmed that toleration is present when describing the armada pod.

Story: 2009232
Task: 43482
Change-Id: I7ec3a3753b49099279450d6228cce8915bc77cd2