Add iproute-tc to the ISO image
iproute was updated from 4.11 to 5.12 with commit a7760b40a266
("iproute: Update from 4.11.0-14 to 5.9.0-4") and commit 36673774ee3c
("iproute-5.12, iptables-1.8.4, and libnftnl-1.1.5"). In CentOS 7, the
iproute-4.11 RPM package included the 'tc' utility, but with CentOS 8,
tc was moved to a separate RPM package, "iproute-tc".

The original version of the second aforementioned commit added the
iproute-tc RPM package in the ISO, but iproute-tc was later removed from
the ISO during the code review, as it was not known at the time that tc
had been split into a separate package (iproute-tc) between CentOS 7's
iproute 4.11 and CentOS 8's 5.9/5.12, and that tc used to be included in
the StarlingX ISO before the iproute updates.

This commit explicitly adds the iproute-tc package into the ISO image to
make tc available in StarlingX installations again.

To prevent another similar issue, the iproute spec file was scanned for
%package declarations to see if any other new sub-packages had been
introduced since CentOS 7, and only the iproute-tc package was seen.

Verification:
- ISO build was successful.
- Resulting bootimage.iso file was confirmed to contain the iproute-tc
  RPM package via "iso-info -f -i bootimage.iso | grep iproute-tc".
- Note that run-time testing of the tc utility had been carried out in
  the past for eBPF testing by manually installing the iproute-tc RPM
  package to a running StarlingX installation.

Closes-Bug: #1951989

Signed-off-by: M. Vefa Bicakci <vefa.bicakci@windriver.com>
Change-Id: I8422f0aa872101e85502f1f3f0d3dc62ec19b8bb