iptables: Include legacy tools in StarlingX
This commit adds a patch to the iptables packaging so that
iptables-legacy and ipt6tables-legacy are included in StarlingX.

This is required in case a container sets up firewall rules with
iptables-legacy and it is necessary to inspect or manipulate these rules
from the host side.

With this patch, /usr/sbin on a StarlingX host will include the
iptables-legacy and ip6tables-legacy symlinks to xtables-legacy-multi.

Verification:
- Incremental monolithic ISO build is successful.
- iptables RPM included in the just-built ISO is confirmed to include
  iptables-legacy and ip6tables-legacy.
- Installation and bootstrap with a VM in all-in-one simplex
  configuration is successful.
- iptables-legacy and ip6tables-legacy binaries were confirmed to be
  able to list firewall rules, as a sanity test.

Partial-Bug: #1953232
Change-Id: I65fa4280d55a16db8c49fb392203033c5e564cdd
Signed-off-by: M. Vefa Bicakci <vefa.bicakci@windriver.com>