debian: Add and patch dh-python package
This is a transient work that enables us to progress Debian
integration. This work affects only Debian.

Sysinv loads stevedore plugins. The logic is based on the fact that
the source files are listed in SOURCES.txt which is included in the
egg. The package that provides a plugin is determined by parsing
entry_point and distribution data. Distribution data loads
information from SOURCES.txt.

dh-python is used during build and silently removes the file that
sysinv expects to be present.

Patch the package to keep the file.

Skip CentOS tests because this is a Debian only package.
Debian tests:
PASS: build python package, build non-package python
PASS: SOURCES.txt present in .egg-info in python package

Partial-Bug: https://bugs.launchpad.net/starlingx/+bug/1955648
Signed-off-by: Dan Voiculeasa <dan.voiculeasa@windriver.com>
Change-Id: Id45b1ba93f9ee88cc01a9cb1fb25527b8a5f1e19