debian: Add missing /etc/etcd directory
When the node is bootstrapped via the ansible playbooks, the
client certs are stored in the /etc/etcd directory, which is
not created by the debian package. This causes the ansible
playbook to fail.

Create the /etc/etcd directory in the etc-server.postinst
script when the server is installed.

Testing:
PASS: Package installed and ISO built successfully

Story: 2009221
Task: 43521

Signed-off-by: Charles Short <charles.short@windriver.com>
Change-Id: I10fbf0bc34a0d65ec75ef9237f12cf63ce1b8c50