Enable generation of Ceph's Python 3 packages
Changed ceph.spec to enable the generation of python 3 packages.
It's important to highlight that the python 2 packages will continue
to be generated and they are the ones used on StarlingX installation.

The python 3 packages will only be, originally on stx-base-image.

There is also a clean up on centos_tarball-dl.lst of commented lines
of ceph submodules that were updated.

Test plan:
Complete build run
Starlingx installation
stx-openstack apply - check that the helm chart can create ceph pools

Depends-On: https://review.opendev.org/c/starlingx/tools/+/824575
Story: 2009074
Task: 44281

Signed-off-by: Delfino Curado <delfinogomes.curadofilho@windriver.com>
Change-Id: I52dac30849a7072b80cad388b16d2b50ea22391a