debian: Correct helm-upload path
helm-upload was located on /usr/sbin while the puppet
module was expecting it to be located in /usr/local/sbin.
Put it in the place where the puppet module was expecting it.

Test Plan:
Pass Build helm with correct paths.
Pass Check to see if helm-upload is located in /usr/local/sbin.

Story: 2009101
Task: 43715

Signed-off-by: Charles Short <charles.short@windriver.com>
Change-Id: I0dc22b4e5888cbfd26728588d0c09a5a47bd3305