Add bond-cni container network plugin
This commit adds the bonding CNI plugin to StarlingX.

The bonding CNI plugin allows a container to bond multiple
interfaces together to be used in a fail-over or load
balancing configuration.

https://github.com/k8snetworkplumbingwg/bond-cni

Note that this plugin (for now) resides outside of the
main containernetwork project, and is still part of the
k8s network plumbing working group project.  As such,
it is required to build this plugin separately.

v1.0 of the bond-cni was released in 2018.  Since then,
14 commits containing such things as doc clean-ups and
bug fixes have been committed.  We pick up these additional
fixes by clamping down on the latest commit SHA.

Testing:

- Configure bond interface name (ifName)
- Configure miimon value (miimon)
- Configure and verify traffic path for modes (mode):
  - balance-rr (0)
  - active-backup (1)
  - balance-xor (2)
  - broadcast (3)
  - 802.3ad (4)
  - balance-tlb (5)
  - balance-alb (6)
- Configure and verify behaviour for fail-over-mac
  modes (failOverMac):
  - none (0)
  - active (1)
  - follow (2)
- Configure linksInContainer:
  - take lower interfaces existing on host
  - take lower interfaces existing on container
- Links tested:
  - virtual interfaces
  - SR-IOV VF interfaces

Story: 2009800
Task: 44344

Change-Id: I7bffaa272ffe9eba85c3aa0a26b9c4f61428b640
Signed-off-by: Steven Webster <steven.webster@windriver.com>