Fix ceph-mon resize on AIO-DX
The system command for resizing ceph-mon partition is disabled
for AIO-DX setup because it is broken.

The problem is that DRBD filesystem is not handled correctly
when the ceph-mon lvm partition is resized.

Updating the DRBD metadata for the ceph-mon partition with
'drbdmadm apply-al' command fixed the problem.

Test-Plan: Install the system with ceph backend and try to resize
           the ceph-mon partition with 'system ceph-mon-modify'
           command.
    PASS: Fresh install and ceph-mon resizing on AIO-SX
    PASS: Fresh install and ceph-mon resizing on AIO-DX

Story: 2009861
Task: 44526

Signed-off-by: Felipe Sanches Zanoni <Felipe.SanchesZanoni@windriver.com>
Change-Id: I734238c94575c94ea59937b72471de940735fe1b