debian: Copy /var/log directory to LVM volume.
When the image boots the LVM volume is mounted via /etc/fstab.
However it does not contain the log directories that services like
Keystone and Apache expect to be present in order for them to start.

This has a knock on affect later in bootstrap when Keystone tries
to restart the service when Keystone is started and it fails to
do so because /var/log/keystone is missing.

Test Plan:
PASS Apply patch
PASS Boot image check for /var/log/keystone.

Story: 2009221
Task: 43324

Signed-off-by: Charles Short <charles.short@windriver.com>
Change-Id: Ib716e48fba2fcf5896ca3f06fc2bae76d24aa05f