Set provider mode when formatting network interfaces file on Debian
When the interfaces file is generated it is necessary to determine the
mode (vlan or raw) from the interface name. Originally the plugin
considers that a previous interfaces file exists and will use it to
set the mode value.

But on StarlingX network_config is generated by sysinv and there is no
previous file to parse, so this change adds the mode logic on
file formatting.

Test Plan:
PASS: generate valid ifupdown config when interface name is of format
      "vlanNNN" on Debian and apply it with ifup/ifdown command

Story: 2009101
Task: 44659

Signed-off-by: Andre Fernando Zanella Kantek <AndreFernandoZanella.Kantek@windriver.com>
Change-Id: Ibccd189ea14eed4b8504908188fabedfa6bf4c2a