Address unmet dependencies in puppet-rabbitmq-8.5.0
The following dependencies were generating
"unresolved dependency" warnings during bootstrap on Debian:

puppet-archive (>= 2.0.0 < 4.0.0)
camptocamp-systemd (>= 2.1.0 < 3.0.0)

But they are not present in Debian Bullseye and not
installable using the base-bullseye.lst.
Also, puppet-archive is not needed since rabbitmq is not
configured to install pkgs at runtime, and camptocamp-systemd
only manages the file limit before bootstrap (sets to 524288)
and removing it does not present big risks.
The solution was to create a patch to remove the
dependency declaration for each module.

Debian Bullseye tests:
PASS: Build & install
PASS: Successful Bootstrap

Story: 2009101
Task: 44710

Signed-off-by: Matheus Machado Guilhermino <Matheus.MachadoGuilhermino@windriver.com>
Change-Id: Ia6025c56dca3ee7f34058057bdd64a6232d30988