Address unmet dependency in puppet-etcd-1.12.3
The following dependency was generating a
"unresolved dependency" warning during bootstrap on Debian:

puppet-stdlib (>= 4.6.0 < 5.0.0)

Debian is currently using puppet-stdlib 5.0.0 and the
most recent version of puppet-etcd asks for puppet-stdlib
(>= 4.6.0 < 5.0.0).
puppet-stdlib 5.0.0 provides the same manifests, classes
and functions as puppet-stdlib 4.25.1, the changes are more
present in the packaging architecture, inclusion of a new function,
and improved and updated documentation.
Updating the dependency declaration to allow version 5.0.0
of puppet-stdlib removes the warning, shows no signs of malfunction,
and does not prevent from successfully bootstraping and unlocking.
For that, the package puppet-etcd-1.12.3 was included and patched.

Debian Bullseye tests:
PASS: Build & install
PASS: Successful Bootstrap
PASS: Successful Unlock

Story: 2009101
Task: 44772

Signed-off-by: Matheus Machado Guilhermino <Matheus.MachadoGuilhermino@windriver.com>
Change-Id: I940faf6c406fa0146da86976154f4e4793621c40