Remove mlx4-config from required Debian services
The mlx4-config.service was deprecated in the change
https://review.opendev.org/c/starlingx/config-files/+/832262 as part of
the Mellanox CX6 support tasks.

This change removes mlx4-config from the list that is being used to keep
track of the services available on CentOS that should also be present on
Debian.

TESTS: Checked the services running before ansible bootstrap in a CentOS
and a Debian install. mlx4-config service is not present in both, which
is the expected behavior.

Story: 2009878
Task: 44868
Signed-off-by: Antonio Augusto Vilas Boas Teixeira
<antonio.augustovilasboasteixeira@windriver.com>
Change-Id: I98b78954068ba0dfdb66bd6b537623634af8d2c0