debian: Update systemd to 247.3-7
Fix build failure due to Debian 11.3. The failure is caused
by us building 247.3-6, while the the package builder was
trying to install systemd 247.3-7 which caused the packages
to fail.

One of the benefits of this is that the systemd-timesync patch
we carry was able to be dropped because it was included
in the latest update.

Test Plan
PASS Build and Install systemd package
PASS Run build-pkgs -a

Story: 2009221
Task: 43416

Signed-off-by: Charles Short <charles.short@windriver.com>
Change-Id: I8b7252695186d1ad9e2e89445fdf7d231c39f6f2