debian: Make /boot/efi mount as rw
Currently the mount is set to ro in fstab.  This changes
the entry in fstab to be "defaults" which will
make it mount rw on the subsequent reboot.

Why do we need it mounted rw?  Because the grub.cfg file
lives on that filesystem, and we need to change it for
things such as kernel bootargs.

Verification
- install on a virtual machine, check the resulting
fstab entry for otaefi being "defaults"
- ensure that /boot/efi is now mounted rw on the
subsequent reboot

Story: 2009964
Task: 45001

Change-Id: I9bf51b1004dc254b8f33d8fd160e546717e7c9f8
Signed-off-by: Jim Somerville <jim.somerville@windriver.com>