Debian: Disable pci-irq-affinity-agent service
This prevents the pci-irq-affinity-agent from being
enabled. This service is not used by the platform
anymore, migrating to be a stx-openstack container [1].

However, there is still WIP for complete the removal [2],
but the agent is already used as a container, it is not
doing anything on the platform.

1: https://storyboard.openstack.org/#!/story/2009299
2: https://review.opendev.org/c/starlingx/utilities/+/830892

Test Plan:
PASS: bootstrap finished successfully
PASS: unlock finished successfully

Story: 2009965
Task: 45194

Signed-off-by: Hugo Brito <hugo.brito@windriver.com>
Change-Id: I6f89653bcf9904362b1f942720cc486a102b7f6f