Debian: Fix deps on openstacklib, mysql modules
The following dependencies were generating the warning
"module 'openstacklib' has unresolved dependencies"
during bootstrap and unlock on Debian:

puppetlabs-openstacklib (v17.4.0) asks for
puppetlabs-postgresql version >=6.4.0 <7.0.0

puppetlabs-mysql (v8.1.0) asks for
puppetlabs-translate version >= 1.0.0 < 2.0.0

Comparing puppetlabs-postgresql v8.0.0 with v6.10.2: It can be
verified that support for Debian 11 was added on v7.4.0, which is
already out of the specified range.
Other than added functionality and fixes, here are the major changes
between v6.10.2(latest version inside of range) and v8.0.0:
v7.0.0 drops support for SLES 11 and RHEL 5, and bumps minimum Puppet
version to 6.0.0 (We are currently using Puppet 5.5.22, but it should
be noted that the minimal version was bumped up because Puppet 5 was
removed from the test cases and not because there are signs of
malfunction).
v8.0.0 drops support for CentOS 6, Debian 6, and Ubuntu 10, which is not
a problem since we are not using any of those OSs.
In conclusion, any version earlier than v7.4.0 should not be used and
there are no known disadvantages to using v8.0.0 instead of v7.4.0.

puppetlabs-translate v2.0.0 removes support for Debian 7 and bumps up
the minimum Puppet version (both of those are irrelevant here since we
are on Debian 11 and the Puppet version is still inside the range).
All other changes introduced from v2.0.0 to v2.2.0 are added support
and minor fixes.
Therefore, it should be safe to use v2.2.0 without a problem.

Debian Bullseye tests:
PASS: Build & install
PASS: Successful Bootstrap
PASS: Successful Unlock

Story: 2009964
Task: 45496

Signed-off-by: Matheus Machado Guilhermino <Matheus.MachadoGuilhermino@windriver.com>
Change-Id: I73fe64b867026ba38b0db7b0a8b34fed388e4d66