Debian: Patch ping for replies with wrong source
This commit patches iputils-ping in Debian-based StarlingX so that ping
prints out ICMP echo replies with unexpected/"wrong" source IP addresses
as was done with CentOS 7's version of ping and as is done with more
recent versions of the iputils package. As an example, if the user calls
"ping 192.168.1.1" and the ICMP echo responses are sent by 192.168.1.3,
then the version of ping in Debian 11 currently assumes that no
responses have been received. This issue was discovered while running
the eBPF test cases in the kernel's samples/bpf source sub-directory.

This issue was introduced upstream with the following commit:
  5e052ada96c3 ("ping: discard packets with wrong source address")
  https://github.com/iputils/iputils/commit/5e052ada96c3

The following commit fixes this issue:
  5f6bec5ab57c ("ping: Print reply with wrong source with warning")
  https://github.com/iputils/iputils/commit/5f6bec5ab57c

Due to a number of contextual dependencies, we unfortunately cannot
directly cherry-pick the commit that fixes this issue. Hence, we
cherry-pick four patches in total. We did not opt for completely
uprevisioning the iputils package to be able to inherit bug fixes that
Debian may introduce in the future.

We checked the list of commits between the two tags 20210202 and
20211215 using the following command:

  git log --oneline refs/tags/20210202..origin/master -- 'ping/*.[ch]'

And we did not find any commits that fix issues in the commits that we
are cherry-picking with this commit.

Testing:
- An ISO image with the standard kernel was successfully incrementally
  built with this commit.
- The resulting ISO image was installed onto a qemu/KVM-based virtual
  machine and the resulting system was bootstrapped with Ansible and
  unlocked.
- The test cases shipped in the kernel repository's samples/bpf
  sub-directory were successfully executed in the virtual machine. These
  test cases make use of ping in IPv4 and IPv6 mode, which ought to be a
  good sanity test.

Partial-Bug: 1977849

Change-Id: I25b42b8de80deec86be0eb504442992863a8d833
Signed-off-by: M. Vefa Bicakci <vefa.bicakci@windriver.com>