debian: Fix containerd shim v2 shutdown
Containerd uses the "process" killmode which shutsdown
shim v1 container processes but it does not shutdown
v2 container processes. As a result, when shutting down
the server it will result in a longer shutdown time
than compared to Centos 7.

This is a temporary workaround until we update to containerd
1.5+.

Test Plan
PASS Build containerd with patch
PASS Boot and unlock server
PASS Reboot server check for continerd-shim processes
in the syslog after the server reboots

Story: 2009845
Task: 44456

Signed-off-by: Charles Short <charles.short@windriver.com>
Change-Id: Iac496e9f2b7f3ccded5ea3e034db8bac2cfc0125