Debian: process route metric parameter
It was detected that the static route's metric parameter is not
applied on the kernel if configured on sysinv database. The cause is
located on the puppet-network module not adding the information to
/var/run/network-scripts.puppet/routes.

This change adds the necessary modification to process the options
field on the hiera file.

Test Plan:
PASS  set parameter with system host-route-add and lock/unlock and
      verify metric value on the kernel

Closes-Bug: 1977983

Signed-off-by: Andre Fernando Zanella Kantek <AndreFernandoZanella.Kantek@windriver.com>
Change-Id: Iafeb2f1d8f6280c84d66398c1d21dbbfe1869a09