debian: fix issue with bond-cni binary install
This commit fixes an issue with the debian install of the
bond CNI plugin.

It was noticed that the 'bond' source directory was being
installed, rather than the built 'bond' binary.

Since the build output is a single binary, it will be
found in debian/bond-cni/, rather than the standard
debian/tmp/

dh_install is instructed where to find the build output
installed by dh_auto_install.

Testing:

- Debian build and install
- Confirmed the bond binary is at /var/opt/cni/bin/
- Functional testing to ensure a pod using the
  bond plugin could be launched

Depends-On: https://review.opendev.org/c/starlingx/integ/+/844865
Closes-Bug: 1976111

Signed-off-by: Steven Webster <steven.webster@windriver.com>
Change-Id: I2a504b8bfc210dc09487a496959da235dec82525