Debian: Wait for devices to be configured after boot
This work is part of AIO-DX integration effort.
This will allow controller-1 to be installed.

When pxebooting controller-1 from controller-0 it is observed that a
dhcp request is initiated before the NIC is renamed. This will lead
to bad network configuration generated during kickstarts.

Wait for NICs to be renamed before doing the dhcp request.

Tests:
PASS: AIO-DX controller-1 gets installed

Story: 2009968
Task: 45826
Signed-off-by: Dan Voiculeasa <dan.voiculeasa@windriver.com>
Change-Id: Ib375d846961e6a04f430293c5e218e31e7afb250