Debian: Break hardlinks in /var
Here the root filesystem is managed by OSTree. OSTree is designed to
manage/deploy read-only filesystem. For read-only filesystem, hardlink
all files with same content is safe and can save disk space. It is a
good principle for readonly folders like /bin, /lib and so on. While
"/var" is somehow special that it is designed to store variable data
files, it can not be read-only, it must be read-write. Files in
"/var" should not be hardlinked together like other folders, break
them to make it.

Do the same for /usr/homedirs/home as it gets mounted at /var/home
at runtime.

Test Plan:
Pass: build ISO image
Pass: In runtime system, no hardlink in /var

Story: 2008862
Task: 45966
Related-Bug: #1983580

Change-Id: I1a61f919e0a60f83e27b6c0bf402c20c2d43abb4
Signed-off-by: Zhang Xiao <xiao.zhang@windriver.com>