Add keystone to /etc/group and /etc/passwd
Add keystone to group, passwd to prevent issue on
data-migration because keystone doesn't have permission
to open the /var/log/keystone/keystone.log and the
content of /opt/platform/keystone/fernet-keys/ as well
on upgrade between 22.06 centos and 22.12 debian.

The fernet-keys content files is created with an uid and
gid specific (42424), without this change the keystone will
not be the owner of the file, with that causing the permission
error. So adding the keystone to these files the user/group
will be keystone keystone instead of 42424 42424.
We want to get the same behavior as we do in CentOS, in [1]

[1]: https://opendev.org/starlingx/integ/src/branch/master/base/setup/centos/patches/0001-Change-group-passwd-and-uidgid.patch

Test Plan:

PASS: Data-migration-complete CentOS -> Debian upgrade
PASS: Debian building test
PASS: Debian AIO-SX unlocked
PASS: Debian AIO-DX unlocked

Story: 2009303
Task: 46093

Signed-off-by: Luis Eduardo Bonatti <LuizEduardo.Bonatti@windriver.com>
Change-Id: Id9bfe914cd80552cb49029c3fdca77886c432a3e